+++ doc/usage/handlers.clj
@@ -26,12 +26,12 @@
 
 ;; Synchronous ring handler - explicitly
 
-(run {:handler (handler/sync-ring-handler ring-handler)})
+(run {:handler (handler/ring-sync ring-handler)})
 ;=> "Hello from localhost (sync)"
 
 ;; Asynchronous ring handler - explicitly
 
-(run {:handler (handler/async-ring-handler ring-handler)})
+(run {:handler (handler/ring-async ring-handler)})
 ;=> "Hello from localhost (async)"
 
 
@@ -39,12 +39,12 @@
 
 ;; Synchronous ring handler - adapter in configuration
 
-(run {:handler ring-handler, :handler-fn-adapter handler/sync-ring-handler})
+(run {:handler ring-handler, :handler-fn-adapter handler/ring-sync})
 ;=> "Hello from localhost (sync)"
 
 ;; Asynchronous ring handler - adapter in configuration
 
-(run {:handler ring-handler, :handler-fn-adapter handler/async-ring-handler})
+(run {:handler ring-handler, :handler-fn-adapter handler/ring-async})
 ;=> "Hello from localhost (async)"
 
 
@@ -52,12 +52,12 @@
 
 ;; Synchronous ring handler - global assignment
 
-(do (server/set-handler-fn-adapter handler/sync-ring-handler)
+(do (server/set-handler-fn-adapter handler/ring-sync)
     (run {:handler ring-handler}))
 ;=> "Hello from localhost (sync)"
 
 ;; Asynchronous ring handler - global assignment
 
-(do (server/set-handler-fn-adapter handler/async-ring-handler)
+(do (server/set-handler-fn-adapter handler/ring-async)
     (run {:handler ring-handler}))
 ;=> "Hello from localhost (async)"